Exercises are started always from scrach, start dialog improved, preparing to storing non guitar instrument scale in Ttune

+++ src/exam/tstartexamdlg.cpp
@@ -44,7 +44,8 @@ TstartExamDlg::TstartExamDlg(QString& nick, QString &path, TexamParams *examPara
     QDialog(parent),
     m_Acction(e_none),
     m_examParams(examParams),
-    m_path(path)
+    m_path(path),
+    m_selectedExamFile("")
 {
     setWindowTitle(tr("Start exercises or an exam"));
 
@@ -76,20 +77,22 @@ TstartExamDlg::TstartExamDlg(QString& nick, QString &path, TexamParams *examPara
 			newLevelLab->setAlignment(Qt::AlignCenter);
 			newLevelLab->setStatusTip(m_createLevelButt->statusTip());
 
-		m_newExerciseButt = new QPushButton(tr("start new exercise"));
-			m_newExerciseButt->setIcon(QIcon(path + "picts/practice.png"));
-			setIconSize(m_newExerciseButt);
-		m_contExerciseButt = new QPushButton(tr("continue exercise"));
-			m_contExerciseButt->setIcon(QIcon(path + "picts/practice.png"));
-			setIconSize(m_contExerciseButt);
+		m_exerciseButt = new QPushButton(tr("exercises"));
+			m_exerciseButt->setIcon(QIcon(path + "picts/practice.png"));
+			setIconSize(m_exerciseButt);
 		
+		m_examMenu = new QMenu("open exam file", this);
 		m_newExamButt = new QPushButton(tr("pass new exam"), this);
 			m_newExamButt->setIcon(QIcon(path + "picts/nootka-exam.png"));
 			setIconSize(m_newExamButt);
 		m_contExamButt = new QPushButton(tr("continue exam"), this);
 			m_contExamButt->setIcon(QIcon(path + "picts/nootka-exam.png"));
 			setIconSize(m_contExamButt);
-    m_examCombo = new QComboBox(this);
+			m_contExamButt->setStatusTip(tr("Click and select an exam to continue"));
+			m_contExamButt->setMenu(m_examMenu);
+		m_lastExamButt = new QPushButton(QIcon(path + "picts/nootka-exam.png"), "", this);
+			setIconSize(m_lastExamButt);
+			m_lastExamButt->setSizePolicy(QSizePolicy::Fixed, QSizePolicy::Fixed);
     
 		m_helpButt = new QPushButton(this);
 			m_helpButt->setIcon(QIcon(path + "picts/help.png"));
@@ -115,24 +118,21 @@ TstartExamDlg::TstartExamDlg(QString& nick, QString &path, TexamParams *examPara
 				helpLevelLay->addWidget(m_createLevelButt);
 				helpLevelLay->addWidget(newLevelLab);
 				helpLevelLay->addStretch(1);
+				helpLevelLay->addWidget(m_cancelBut);
 			upperLay->addLayout(helpLevelLay);
 		QVBoxLayout *mainLay = new QVBoxLayout;
 			mainLay->addLayout(upperLay);
-			QHBoxLayout *exerciseButtLay = new QHBoxLayout;
-				exerciseButtLay->addWidget(m_newExerciseButt);
-				exerciseButtLay->addWidget(m_contExerciseButt);
-			mainLay->addLayout(exerciseButtLay);
+			QHBoxLayout *exerciseLay = new QHBoxLayout;
+				exerciseLay->addSpacing(40);
+				exerciseLay->addWidget(m_exerciseButt);
+				exerciseLay->addSpacing(40);
+			mainLay->addLayout(exerciseLay);
 		QHBoxLayout *examLay = new QHBoxLayout;
 			examLay->addWidget(m_newExamButt);
-			QVBoxLayout *contExamLay = new QVBoxLayout;
-				contExamLay->addWidget(m_contExamButt);
-				contExamLay->addWidget(m_examCombo);
-			examLay->addLayout(contExamLay);
+			examLay->addWidget(m_contExamButt);
+			examLay->addWidget(m_lastExamButt);
 		mainLay->addLayout(examLay);
-			QHBoxLayout *bottomLay = new QHBoxLayout;
-				bottomLay->addWidget(m_hintLabel);
-				bottomLay->addWidget(m_cancelBut);
-		mainLay->addLayout(bottomLay);
+		mainLay->addWidget(m_hintLabel);
 			
     setLayout(mainLay);
     
@@ -144,51 +144,53 @@ TstartExamDlg::TstartExamDlg(QString& nick, QString &path, TexamParams *examPara
 #else
     QSettings sett;
 #endif
+		QAction *loadExamAct = new QAction(loadExamFileTxt(), this);
+			loadExamAct->setIcon(QIcon(path + "picts/nootka-exam.png"));
+			connect(loadExamAct, SIGNAL(triggered(bool)), this, SLOT(examFromFileDialog()));
+		m_examMenu->addAction(loadExamAct);
+		m_examMenu->addSeparator();
+		m_examMenu->addAction(tr("recent opened exams:"));
+		
     m_recentExams = sett.value("recentExams").toStringList();
-    for (int i = m_recentExams.size()-1; i >= 0; i--) {
+    for (int i = 0; i < m_recentExams.size(); i++) {
         QFileInfo fi(m_recentExams[i]);
         if (fi.exists()) {
-            m_examCombo->insertItem(0, m_recentExams[i]);
+						QAction *recentAct = m_examMenu->addAction(m_recentExams[i]);
+						connect(recentAct, SIGNAL(triggered(bool)), this, SLOT(continuePrevExam()));
         }
         else
             m_recentExams.removeAt(i);
     }
-    m_examCombo->addItem(loadExamFileTxt());
     if (m_recentExams.size()) {
         sett.setValue("recentExams", m_recentExams);
-        m_examCombo->setCurrentIndex(0);
     }
 		
 		updateButtonStatusText("");
 		QString exerciseFile = QDir::toNativeSeparators(QFileInfo(sett.fileName()).absolutePath() + "/exercise.noo");
-		bool continuePosiible = false;
+		m_prevExerciseLevel.name = ""; // empty means - no previous level
 		if (QFileInfo(exerciseFile).exists()) {
-				Tlevel l;
-				Texam exam(&l, "");
+				Texam exam(&m_prevExerciseLevel, "");
         Texam::EerrorType err = exam.loadFromFile(exerciseFile);
-        if (err == Texam::e_file_OK || err == Texam::e_file_corrupted) {
-						continuePosiible = true; // TODO prevExerciseLevel makes no sense - level name is taken from real file
-						m_examParams->prevExerciseLevel = exam.level()->name;
-				} else {
-					qDebug() << "exercise file was corrupted... and deleted...";
-					QFile::remove(exerciseFile);
+        if (err != Texam::e_file_OK && err != Texam::e_file_corrupted) {
+						qDebug() << "exercise file was corrupted... and deleted...";
+						QFile::remove(exerciseFile);
 				}
 		}
-		if (continuePosiible)
-				m_contExerciseButt->setStatusTip(m_contExerciseButt->text() + "<br>" +
-															tr("on level:") + " <b>" + m_examParams->prevExerciseLevel + "</b>");
+		if (m_prevExerciseLevel.name != "")
+			 m_exerciseButt->setStatusTip(tr("Start exercise on level:") + "<br><b>" + m_prevExerciseLevel.name + "</b>");
+		
+		if (m_recentExams.size())
+				m_lastExamButt->setStatusTip(tr("Continue the latest exam:") + "<br><b>" + m_recentExams.at(0) + "</b>");
 		else
-				m_contExerciseButt->setDisabled(true);
-		m_contExamButt->setStatusTip(m_contExamButt->text() + "<br>" + " <b>" + m_examCombo->currentText() + "</b>");
+				m_lastExamButt->hide();
 		
     connect(m_levelsView, SIGNAL(levelToLoad()), this, SLOT(levelToLoad()));
     connect(m_cancelBut, SIGNAL(clicked()), this, SLOT(reject()));
     connect(m_createLevelButt, SIGNAL(clicked()), this, SLOT(createLevel()));
     connect(m_newExamButt, SIGNAL(clicked()), this, SLOT(startAccepted()));
     connect(m_contExamButt, SIGNAL(clicked()), this, SLOT(startAccepted()));
-		connect(m_newExerciseButt, SIGNAL(clicked()), this, SLOT(startAccepted()));
-		connect(m_contExerciseButt, SIGNAL(clicked()), this, SLOT(startAccepted()));
-    connect(m_examCombo, SIGNAL(activated(int)), this, SLOT(prevExamSelected(int)));
+		connect(m_lastExamButt, SIGNAL(clicked()), this, SLOT(continueTheLast()));
+		connect(m_exerciseButt, SIGNAL(clicked()), this, SLOT(startAccepted()));
     connect(m_levelsView, SIGNAL(levelChanged(Tlevel)), this, SLOT(levelWasSelected(Tlevel)));
 		connect(m_helpButt,  SIGNAL(clicked()), this, SLOT(helpSelected()));
 		
@@ -206,26 +208,23 @@ TstartExamDlg::TstartExamDlg(QString& nick, QString &path, TexamParams *examPara
 TstartExamDlg::Eactions TstartExamDlg::showDialog(QString &txt, Tlevel &lev) {
     exec();
     if (result() == QDialog::Accepted) {
-        if (m_Acction == e_newExam || m_Acction == e_newExercise) {
+        if (m_Acction == e_newExam || m_Acction == e_runExercise) {
             txt = m_nameEdit->text();
-            lev = m_levelsView->getSelectedLevel();
+						if (m_prevExerciseLevel.name == "")
+								lev = m_levelsView->getSelectedLevel();
+						else
+								lev = m_prevExerciseLevel;
             return m_Acction;
-        }
-        else if (m_Acction == e_contExam) {
-            if (m_examCombo->currentIndex() != -1) {
-                txt = m_examCombo->currentText();
+        } else if (m_Acction == e_contExam) {
+            if (m_selectedExamFile != "") {
+                txt = m_selectedExamFile;
                 return e_contExam;
             } else
 								return e_none;
-        } else {
-						return e_contExercise;
-        }
-    } else {
-		if (m_Acction == e_levelCreator)
-			return e_levelCreator;
-		else
-			return e_none;
-	}
+					}
+    } else if (m_Acction == e_levelCreator)
+						return e_levelCreator;
+		return e_none;
 }
 
 
@@ -243,6 +242,13 @@ bool TstartExamDlg::event(QEvent *event) {
 }
 
 
+void TstartExamDlg::continueTheLast() {
+		if (m_recentExams.size()) {
+				examToContSelected(m_recentExams[0]);
+		}
+}
+
+
 bool TstartExamDlg::isAnyLevelSelected() {
 	Tlevel l = m_levelsView->getSelectedLevel();
 	if (l.name == "") { // nothing selected
@@ -254,43 +260,29 @@ bool TstartExamDlg::isAnyLevelSelected() {
 
 
 void TstartExamDlg::startAccepted() {
-    if (sender() == m_newExamButt || sender() == m_newExerciseButt) { // new exam or exercise on selected level
+    if (sender() == m_newExamButt) { // new exam on selected level
         if (!isAnyLevelSelected())
             return;
-        else {
-            if (m_nameEdit->text() == "") {
-                QMessageBox::warning(this, "", tr("Give a user name!"));
-                return;
-            }
-            if (sender() == m_newExerciseButt)
-							m_Acction = e_newExercise;
-						else
-							m_Acction = e_newExam;
-            accept();
-        }
-    } else { // exam or exercise to continue
-				if (sender() == m_contExerciseButt) {
-							m_Acction = e_contExercise;
-							accept();
-				} else
-					if (m_examCombo->currentIndex() < m_examCombo->count() - 1)    {
-            m_Acction = e_contExam;
-            accept();
-        }
-    }
-}
-
-
-void TstartExamDlg::loadExam() {
-    QString fileName = QFileDialog::getOpenFileName(this, loadExamFileTxt(),
-                               m_examParams->examsDir, examFilterTxt(), 0, QFileDialog::DontUseNativeDialog);
-    if (fileName != "") {
-        m_examParams->examsDir = QFileInfo(fileName).absoluteDir().absolutePath();
-        m_examCombo->insertItem(0, fileName);
-        m_recentExams.prepend(fileName);
-        m_examCombo->setCurrentIndex(0);
-        m_Acction = e_contExam;
-        accept();
+				if (m_nameEdit->text() == "") {
+						QMessageBox::warning(this, "", tr("Give a user name!"));
+						return;
+				}
+				m_Acction = e_newExam;
+				accept();
+    } else if (sender() == m_exerciseButt) { // exercise on selected level or previous one
+				if (m_prevExerciseLevel.name == "" && !isAnyLevelSelected())
+					return;
+				if (m_nameEdit->text() == "") {
+						QMessageBox::warning(this, "", tr("Give a user name!"));
+						return;
+				}
+				m_Acction = e_runExercise;
+				accept();
+    } else { // exam to continue
+				if (m_selectedExamFile != "")    {
+					m_Acction = e_contExam;
+					accept();
+			}
     }
 }
 
@@ -301,25 +293,44 @@ TstartExamDlg::Eactions TstartExamDlg::createLevel() {
 }
 
 
-void TstartExamDlg::prevExamSelected(int index) {
-    m_contExamButt->setStatusTip(m_contExamButt->text() + "<br><b>" + m_examCombo->currentText() + "</b>");
-    if (index == m_examCombo->count() - 1) {
-        loadExam();
-    } else {
-        m_Acction = e_contExam;
-        accept();
+void TstartExamDlg::continuePrevExam() {
+	QAction *action = qobject_cast<QAction *>(sender());
+	if (action)
+			examToContSelected(action->text());
+}
+
+
+void TstartExamDlg::examFromFileDialog() {
+    QString fileName = QFileDialog::getOpenFileName(this, loadExamFileTxt(),
+                               m_examParams->examsDir, examFilterTxt(), 0, QFileDialog::DontUseNativeDialog);
+    if (fileName != "") {
+				examToContSelected(fileName);
     }
+}
+
+
+void TstartExamDlg::examToContSelected(QString eFile) {
+		if (eFile != "") {
+				m_examParams->examsDir = QFileInfo(eFile).absoluteDir().absolutePath();
+        m_recentExams.prepend(eFile);
+				m_selectedExamFile = eFile;
+				m_Acction = e_contExam;
+        accept();
+		}
 }
 
 
 void TstartExamDlg::updateButtonStatusText(const QString& levelName) {
 	QString statusText;
-	if (levelName == "")
-			statusText += "<br><b>" + tr("No level was selected!") + "</b>";
-	else 
-			statusText += "<br><b>" + levelName + "</b>";
+	if (levelName == "") {
+			statusText += tr("No level was selected!");
+	}	else {
+			statusText += levelName;
+	}
+	statusText.prepend("<br><b>");
+	statusText.append("</b>");
 	m_newExamButt->setStatusTip(tr("Pass new exam on level:") + statusText);
-	m_newExerciseButt->setStatusTip(tr("New exercise on level:") + statusText);
+	m_exerciseButt->setStatusTip(tr("Start exercise on level:") + statusText);
 }
